requirements.txt
pandas==1.3.1
slackclient==2.9.3
click==7.1.2